pyproject.toml
[tool.poetry]
name = "misc"
version = "0.1.0"
description = ""
authors = ["Pai-Sho <paisho@keeperdao.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
matplotlib = "^3.5.1"
requests = "^2.27.1"
numpy = "^1.22.1"
scipy = "^1.7.3"
web3 = "^5.26.0"
bitcoin = "^1.1.42"
websocket-client = "1.0.0"
jsonpickle = "^2.1.0"
pysha3 = "^1.0.2"
pytz = "^2021.3"
python-dotenv = "^0.19.2"
multicall = "^0.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
